Update moving UNet
